Allow independent forcing and mesh. Revise forcing config.
