+++ examples/Greenland_CARRA2/forcing/1990-05-01/P1M.nc
