+++ src/ebfm/core/config/__init__.py
@@ -5,3 +5,4 @@
 from ebfm.core.config.time import TimeConfig, Calendar, DEFAULT_TZ, iso8601  # noqa: F401
 from ebfm.core.config.coupling import CouplingConfig, ComponentId, FieldValidationLevel  # noqa: F401
 from ebfm.core.config.grid import GridConfig, GridInputType  # noqa: F401
+from ebfm.core.config.forcing import ForcingConfig, ForcingType  # noqa: F401
