+++ src/ebfm/main.py
@@ -17,7 +17,7 @@ from ebfm.core import (
 )
 from ebfm.core import LOOP_write_to_file, FINAL_create_restart_file
 from ebfm.core.grid import GridInputType
-from ebfm.core.config import CouplingConfig, GridConfig, TimeConfig
+from ebfm.core.config import CouplingConfig, GridConfig, TimeConfig, ForcingConfig
 from ebfm.core.logger import Logger, setup_logging, log_levels_map, getLogger
 from ebfm.core.cli import (
     extract_active_coupling_features,
@@ -126,12 +126,14 @@ def _main_impl():
         reset_handlers=True,
     )
 
+    logger = getLogger(__name__)
+    logger.info(f"Starting EBFM version {ebfm.core.get_version()}...")
+
+    forcing_config = ForcingConfig(coupling_config, args)
+
     if not hasattr(args, "use_part"):
         # If not provided via command line option --use-part, set to rank + 1 (assuming partition IDs start at 1).
         args.use_part = ebfm_comm.rank + 1
-
-    logger = getLogger(__name__)
-    logger.info(f"Starting EBFM version {ebfm.core.get_version()}...")
 
     # Numba is opt-in: only activate when --with-numba is explicitly passed.
     if args.with_numba:
@@ -223,7 +225,6 @@ def _main_impl():
             IN["Pres"] = data_from_icon["sfcpres"]
 
         # Read/set meteorological forcing
-        forcing_config = LOOP_climate_forcing.ForcingConfig(grid_config, coupler)  # use args?
         IN, OUT = LOOP_climate_forcing.main(C, grid, IN, t, time, OUT, forcing_config)
 
         # Run surface energy balance model
