+++ tox.ini
@@ -49,6 +49,7 @@ extras =
     dev
 commands =
     ebfm --netcdf-mesh examples/Greenland_CARRA2/grid.nc --start-time 1990-05-01T00:00:00 --end-time 1990-05-02T00:00:00
+    ebfm --netcdf-mesh examples/Greenland_CARRA2/grid.nc --forcing-dir examples/Greenland_CARRA2/forcing --start-time 1990-05-01T00:00:00 --end-time 1990-05-02T00:00:00
 
 [testenv:type-check]
 description = Run static type checking
